fix verification workflow on windows
This is just an attempt, but I think the problem was happening because
the verification window would close before any other window would open.
On Windows when the last window is closed, the app quits.

So now it’s tracking whether it’s in the verification workflow or not.
If it is, the app doesn’t quit when all windows are closed.