Make the series lines actions the same as the lines actions
This means that the series lines data is mirrored into the lines data
in the Redux store because both the lines and series lines reducers
will listen for the actions.  I can now make sure that the data looks
equivalent when it arrives in the lines part of the Redux store.