Hack: fix import from URL in card dialog (#512)
I suspect that focus is actually being stolen by the modal which the
image picker model is on top of.  There's a mousedown event which
appears to steal focus (if I delete it the event handler from the
developer tools then focus works as expected and the user can enter
text into the input field).  I recommend that this hack be removed
when we track down where the mousedown event comes from.